+++ src/services/routes.js
@@ -1,25 +1,19 @@
 const fs = require('fs')
-const stripe = rootRequire('services/integrations/stripe')
 const cors = require('cors')
 
 function init(app, express) {
   app.use(cors({ origin: true }))
-  setBodyParser(app, express)
-  setRoutes(app)
-}
-
-function setBodyParser(app, express) {
-  app.use(express.json())
   app.use(express.urlencoded({ extended: true }))
+
+  setRoutes(app, express)
 }
 
-function setRoutes(app) {
+function setRoutes(app, express) {
   fs.readdirSync(srcPath('services/routes')).forEach(route => {
-    rootRequire(`services/routes/${route}`).setRoute(app, stripe)
+    rootRequire(`services/routes/${route}`).setRoute(app, express)
   })
 }
 
 module.exports = {
-  init,
-  stripe
+  init
 }