+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename project.
 - Bump `lodash` version number to `4.17.20`.
 - Rename routes.
+- Refactor routes.
 
 ### Fixed
 - Handling of line items quantity default value.